Tue, Nov 7, 2023, 2:40 PM +05:30

+++ filter-less/helpers.c
@@ -7,10 +7,10 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0 ; j < width ; j++)
         {
-            int n =  (int)(images[i][j].rgbtRed + images[i][j].rgbtGreen + images[i][j].rgbtBlue) / 3.0
-            images[i][j].rgbtRed = n ;
-            images[i][j].rgbtGreen = n ;
-            images[i][j].rgbtBlue = n ;
+            float n =  (images[i][j].rgbtRed + images[i][j].rgbtGreen + images[i][j].rgbtBlue) / 3.0
+            images[i][j].rgbtRed = (int)n ;
+            images[i][j].rgbtGreen =(int) n ;
+            images[i][j].rgbtBlue = (int) n ;
         }
     }
     return;
